Added benchmark tests for value setting as well
