+++ TryAtSoftware.Extensions.Reflection.Benchmark/Program.cs
@@ -1,4 +1,5 @@
 ﻿using BenchmarkDotNet.Running;
 using TryAtSoftware.Extensions.Reflection.Benchmark;
 
-BenchmarkRunner.Run<ReflectionVsExpressions>();
+BenchmarkRunner.Run<ValueRetrieving>();
+BenchmarkRunner.Run<ValueSetting>();
